move spotlight components to common
