+++ src/components/common/Spotlight.tsx
