+++ src/sass/app.scss
@@ -3,7 +3,7 @@
 @import "Layout.scss";
 @import "components/common/Header.scss";
 @import "components/common/EnvironmentBanner.scss";
-@import "components/base/SpotlightAction.scss";
+@import "components/common/SpotlightAction.scss";
 
 @import "components/common/MarkdownText.scss";
 @import "components/common/UserCard.scss";
