+++ src/sass/components/common/SpotlightAction.scss
